+++ packages/pluggableWidgets/bottom-sheet-native/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Fixed
+
+-   Fixed iOS orientation change when opening the bottom sheet modal on devices in landscape mode.
+
 ## [5.3.2] - 2026-8-4
 
 ### Fixed