+++ packages/pluggableWidgets/bottom-sheet-native/CHANGELOG.md
@@ -6,9 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+-   Added an `On change` event that triggers when the `Trigger attribute` value changes, allowing detection of bottom sheet dismissal (e.g. by outside click).
+
 ### Fixed
 
 -   Fixed iOS orientation change when opening the bottom sheet modal on devices in landscape mode.
+-   Replaced a deprecated prop 'containerHeight' with the respective 'containerLayoutState' option.
 
 ## [5.3.2] - 2026-8-4